backend/requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-PyMongo==2.3.0
Flask-Cors==4.0.1
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
pymongo==4.7.3
eventlet==0.36.1
dnspython==2.6.1
